Creation 2021 annual report
A preliminary copy of the 2021 annual report has been added.
The 2020 images haven't been deleted(It will be done on a next commit),
but all the content in the
https://docs.google.com/document/d/1xexx4YBU5fofSZLAPAN5ddtjWyQP556LWyTZVQ7EbcI/edit#heading=h.9xqyqv28ku8k
file has been used and replaced on the 2021 HTML version, and also all
the information that needs an update has been tagged on the HTML.

As you can see in the HTML, some sections have been commented out
instead of erased. This is because the copy still needs to be updated.
Therefore I don't want to get rid of components that potentially could
be used later on.